internal/civisibility: split instrumentation.go in two files with instrumentation_orchestrion.go representing the api exposed for the orchestrion integration.
